requirements.txt
anyio==3.6.2
APScheduler==3.6.3
black==23.3.0
cachetools==4.2.2
certifi==2022.12.7
click==8.1.3
colorama==0.4.6
flake8==6.0.0
h11==0.14.0
httpcore==0.16.3
httpx==0.24.1
idna==3.4
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
pycodestyle==2.10.0
pyflakes==3.0.1
python-dotenv==1.0.0
python-telegram-bot==20.3
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tomli==2.0.1
tornado==6.2
tzdata==2023.3
tzlocal==4.3
